Add motivation and example re. extensibility
I thought it might be useful to have a motivation and example about the
need for library users to be able to define domain specific quantities
and units with all the functionality provided by library quantitites and
units.

I wrote this on the plane after CppCon when I was a bit tired, so maybe
it needs an editing pass for clarity.